test: split into files
