+++ jest.config.ts
@@ -1,0 +1,11 @@
+export default {
+    coverageDirectory: "./artifacts/coverage",
+    coveragePathIgnorePatterns: [
+        "<rootDir>/node_modules",
+        "<rootDir>/tests",
+        "<rootDir>/e2e",
+    ],
+    moduleDirectories: ["./node_modules", "<rootDir>/node_modules", "."],
+    preset: "ts-jest",
+    testPathIgnorePatterns: ["./artifacts/", "./dist/", "./node_modules/"],
+};
