--- jest.e2e.config.json
@@ -1,12 +1,0 @@
-{
-  "preset": "ts-jest",
-  "moduleDirectories": ["./node_modules", "<rootDir>/node_modules", "."],
-  "moduleNameMapper": {
-    "src(.*)": "<rootDir>/dist$1"
-  },
-  "testPathIgnorePatterns": [
-    "./artifacts/",
-    "./dist/",
-    "./node_modules/"
-  ]
-}
