+++ jest.e2e.config.ts
@@ -1,0 +1,8 @@
+import config from "./jest.config";
+
+export default {
+    ...config,
+    moduleNameMapper: {
+        "src(.*)": "<rootDir>/dist$1",
+    },
+};
