+++ tests/utils/combineReducers.test.ts
@@ -1,20 +1,4 @@
-import { createAction } from "src";
-import { bindActionCreator } from "src/utils/bindActionCreators";
 import { combineReducers } from "src/utils/combineReducers";
-
-describe("bindActionCreators", () => {
-    it("binds action creator function to dispatch", () => {
-        const dispatch = jest.fn();
-        const action = { payload: "Some payload", type: "ACTION_TYPE" };
-        const boundActionCreator = bindActionCreator(
-            createAction(action.type, () => action),
-            dispatch,
-        );
-        expect(boundActionCreator!()).toBeUndefined();
-        expect(dispatch).toHaveBeenCalledTimes(1);
-        expect(dispatch).toHaveBeenCalledWith(action);
-    });
-});
 
 describe("combineReducers", () => {
     it("calls all reducers", () => {
